QEDprocesses

Documentation for QEDprocesses.

QEDprocesses.AbstractComputationSetup

QEDprocesses.AbstractInvalidInputException

QEDprocesses.AbstractModelDefinition

QEDprocesses.AbstractProcessDefinition

QEDprocesses.AbstractProcessSetup

QEDprocesses.InvalidInputError

QEDprocesses._assert_valid_input

QEDprocesses._base_component_type

QEDprocesses._compute

QEDprocesses._differential_cross_section

QEDprocesses._post_processing

QEDprocesses._total_cross_section

QEDprocesses.compute

QEDprocesses.differential_cross_section

QEDprocesses.fundamental_interaction_type

QEDprocesses.incoming_particles

QEDprocesses.number_incoming_particles

QEDprocesses.number_outgoing_particles

QEDprocesses.outgoing_particles

QEDprocesses.physical_model

QEDprocesses.scattering_process

QEDprocesses.total_cross_section

QEDprocesses.AbstractComputationSetup — Type

Abstract base type for computation setups. A setup means a collection of setup data needed to evaluate a dedicated quantity of given running data. Therefore, each setup is associated with a single quantity, which one may compute using the setup data and the running data. Despite that, the decomposition into setup and running data is arbitrary, and this can be used for cases where a subset of the variables a quantity depends on is kept constant.

Computation setup interface

The computation performed using a computation setup is separated into three steps:

1. input validation
2. actual computation
3. post processing

where every step has its own interface function (see compute for details).

Input validation

Every subtype of AbstractComputationSetup should implement the interface function

_assert_valid_input(stp::AbstractComputationSetup, input)

which should throw and an exception subtyped from AbstractInvalidInputException if the input is not valid for the computation of the associated quantity (see _assert_valid_input for more details). The default implementation does nothing, i.e. every input is valid by default. Provide a custom implementation if a different behavior is required.

Actual computation

Every subtype of AbstractComputationSetup must at least implement the required interface function

_compute(stp::AbstractComputationSetup, input) 

which computes the value of the associated quantity for a given input (see _compute for more details).

Post processing

Every subtype of AbstractComputationSetup should implement the interface function

_post_processing(stp::AbstractComputationSetup, input, result) 

which performs task after the actual computation, e.g. conversions or normalizations (see _post_processing for more details).

QEDprocesses.AbstractInvalidInputException — Type

Abstract base type for exceptions indicating invalid input. See InvalidInputError for a simple concrete implementation. Concrete implementations should at least implement

Base.showerror(io::IO, err::CustomInvalidError) where {CustomInvalidError<:AbstractInvalidInputException}

QEDprocesses.AbstractModelDefinition — Type

Abstract base type for all compute model definitions in the context of scattering processes. Every subtype of AbstractModelDefinition is associated with a fundamental interaction. Therefore, one needs to implement the following soft interface function

fundamental_interaction_type(::AbstractModelDefinition)

QEDprocesses.AbstractProcessDefinition — Type

Abstract base type for definitions of scattering processes. It is the root type for the process interface, which assumes that every subtype of AbstractProcessDefinition implements at least

incoming_particles(proc_def::AbstractProcessDefinition)
outgoing_particles(proc_def::AbstractProcessDefinition)

which return a tuple of the incoming and outgoing particles, respectively.

QEDprocesses.AbstractProcessSetup — Type

Abstract base type for setups related to combining scattering processes and physical models. Every subtype of AbstractProcessSetup must implement at least the following interface functions:

scattering_process(::AbstractProcessSetup) 
physical_model(::AbstractProcessSetup) 

Derived from these interface functions, the following delegations are provided:

number_incoming_particles(::AbstractProcessSetup)
number_outgoing_particles(::AbstractProcessSetup)

QEDprocesses.InvalidInputError — Type

InvalidInputError(msg::String)

Exception which is thrown if a given input is invalid, e.g. passed to _assert_valid_input.

QEDprocesses._assert_valid_input — Method

_assert_valid_input(stp::AbstractComputationSetup, input::Any)

Interface function, which asserts that the given input is valid, and throws an InvalidInputError if not.

default implementation

By default, every input is assumed to be valid. Therefore, this function does nothing. To customize this behavior, add your own implementation of

_assert_valid_input(stp::YourCustomSetup,input)

which should throw an exception, which is a subtype of AbstractInvalidInputException. One may also use the concrete implementation InvalidInputError if the input is invalid instead of writing a custom exception type.

QEDprocesses._base_component_type — Method

_base_component_type(array_of_lv::AbstractArray{LV}) where {LV<:QEDbase.AbstractLorentzVector}

Return the type of the components of given Lorentz vectors, which are by themself elements of an AbstractArray.

Examples

julia> using QEDbase
julia> using QEDprocesses
julia> v = Vector{SFourMomentum}(undef,10)
julia> QEDprocesses._base_component_type(v)
Float64

QEDprocesses._compute — Function

_compute(stp::AbstractComputationSetup, input::Any)

Interface function that returns the value of the associated quantity evaluated on input, which can be anything the associated quantity is defined to be feasible for.

unsafe implementation

This function must be implemented for any subtype of AbstractComputationSetup. It should not do any input validation or post processing (see _assert_valid_input and _post_processing), as those two are performed while calling the safe version of this function compute.

QEDprocesses._differential_cross_section — Function

_differential_cross_section(
    proc_def::AbstractProcessDefinition,
    model_def::AbstractModelDefinition,
    in_phasespace::AbstractVector{T},
    out_phasespace::AbstractVector{T},
) where {T<:QEDbase.AbstractFourMomentum}

Interface function for the combination of scattering processes and physical models. Return the differential cross section of a given process and physical model for a passed initial and final phase space. The elements of the AbstractVector representing the phase spaces are the momenta of the respective particles. The implementation of this function for a concrete process and model must not check if the length of the passed phase spaces match the respective number of particles.

differential cross section interface

Given an implementation of this method, the following unsafe generic implementations are provided:

_differential_cross_section(proc_def,model_def,in_phasespace::AbstractVector{T},finial_phasespace::AbstractMatrix{T})
_differential_cross_section(proc_def,model_def,in_phasespace::AbstractMatrix{T},finial_phasespace::AbstractVector{T})
_differential_cross_section(proc_def,model_def,in_phasespace::AbstractMatrix{T},finial_phasespace::AbstractMatrix{T})

where T<:QEDbase.AbstractFourMomentum. Although, any combinations of initial and final phase space types given by single set of points (AbstractVector{T}) and mutiple set of points (AbstractMatrix{T}) is implemented. Furthermore, a safe version of _differential_cross_section is also implemented: differential_cross_section.

unsafe implementation

Each instance of this function does not check the validity of the input. Therefore, these functions are not exported and should be used with caution. To add a method in order to implement the cross section interface, it is recommented to directly use QEDprocesses._differential_cross_section instead of globally using QEDprocesses: _differential_cross_section.

QEDprocesses._post_processing — Method

function _post_processing(stp::AbstractComputationSetup, input::Any, result::Any)

Interface function, which is called in compute after _compute has been called. This function is dedicated to finalize the result of a computation.

default implementation

Since in the case of no post processing the result of _compute is unchanged, this function returns result by default.

QEDprocesses._total_cross_section — Function

_total_cross_section(
    proc_def::AbstractProcessDefinition,
    model_def::AbstractModelDefinition,
    in_phasespace::AbstractVector{T},
) where {T<:QEDbase.AbstractFourMomentum} end

Interface function for the combination of scattering processes and physical models. Return the total cross section of a given process and model for a passed initial phase space. The elements of the AbstractVector representing the initial phase space are the momenta of the respective particles. The implementation of this function for a concrete process and model must not check if the length of the passed initial phase spaces match number of incoming particles.

cross section interface

Given an implementation of this method, the following unsafe generic implementation is provided:

_total_cross_section(proc_def,model_def,in_phasespace::AbstractMatrix{T})

where T<:QEDbase.AbstractFourMomentum. Although, _total_cross_section is also implemented for a vector of initial phase space points. Furthermore, a safe version of _total_cross_section is also implemented: total_cross_section.

Note

Each instance of this function does not check the validity of the input. This function is not exported and should be used with caution. To add a method in order to implement the cross section interface, it is recommented to directly use QEDprocesses._total_cross_section instead of globally using QEDprocesses: _total_cross_section.

QEDprocesses.compute — Method

compute(stp::AbstractComputationSetup, input::Any)

Return the value of the quantity associated with stp for a given input. In addition to the actual call of the associated unsafe version _compute, input validation ([_assert_valid_input]) and post processing (using _post_processing) are wrapped around the calculation (see AbstractComputationSetup for details).

QEDprocesses.differential_cross_section — Method

differential_cross_section(
    proc_def::AbstractProcessDefinition,
    model_def::AbstractModelDefinition,
    in_phasespace::Union{AbstractVector{T},AbstractMatrix{T}},
    out_phasespace::Union{AbstractVector{T},AbstractMatrix{T}},
) where {T<:QEDbase.AbstractFourMomentum}

Return the differential cross section for a given combination of a scattering process and model definition evaluated on the passed inital and final phase space points.

This function will eventually call the respective interface function _differential_cross_section.

QEDprocesses.fundamental_interaction_type — Function

fundamental_interaction_type(models_def::AbstractModelDefinition)

Return the fundamental interaction associated with the passed model definition.

QEDprocesses.incoming_particles — Function

incoming_particles(proc_def::AbstractProcessDefinition)

Interface function for scattering processes. Return a tuple of the incoming particles for the given process definition. This function needs to be given to implement the scattering process interface.

QEDprocesses.number_incoming_particles — Method

number_incoming_particles(
    proc_def::AbstractProcessDefinition
)

Return the number of incoming particles of a given process.

QEDprocesses.number_outgoing_particles — Method

number_outgoing_particles(
    proc_def::AbstractProcessDefinition
)

Return the number of outgoing particles of a given process.

QEDprocesses.outgoing_particles — Function

outgoing_particles(proc_def::AbstractProcessDefinition)

Interface function for scattering processes. Return the tuple of outgoing particles for the given process definition. This function needs to be given to implement the scattering process interface.

QEDprocesses.physical_model — Function

physical_model(stp::AbstractProcessSetup)

Interface function that returns the physical model associated with stp, i.e. an object which is a subtype of AbstractModelDefinition.

QEDprocesses.scattering_process — Function

scattering_process(stp::AbstractProcessSetup)

Interface function that returns the scattering process associated with stp, i.e. an object which is a subtype of AbstractProcessDefinition.

QEDprocesses.total_cross_section — Method

total_cross_section(
    proc_def::AbstractProcessDefinition,
    model_def::AbstractModelDefinition,
    in_phasespace::Union{AbstractVector{T},AbstractMatrix{T}},
) where {T<:QEDbase.AbstractFourMomentum}

Return the total cross section for a combination of a scattering process and a physical model evaluated on a given initial phase space.

This function will eventually call the respective interface function _total_cross_section.