QEDprocesses

Documentation for QEDprocesses.

QEDprocesses.AbstractComputationSetup

QEDprocesses.AbstractInvalidInputException

QEDprocesses.AbstractModelDefinition

QEDprocesses.AbstractProcessDefinition

QEDprocesses.AbstractProcessSetup

QEDprocesses.InvalidInputError

QEDprocesses.PhasespaceDefinition

QEDprocesses._assert_valid_input

QEDprocesses._averaging_norm

QEDprocesses._base_component_type

QEDprocesses._compute

QEDprocesses._generate_incoming_momenta

QEDprocesses._generate_momenta

QEDprocesses._generate_outgoing_momenta

QEDprocesses._incident_flux

QEDprocesses._is_in_phasespace

QEDprocesses._matrix_element

QEDprocesses._phase_space_factor

QEDprocesses._post_processing

QEDprocesses._total_probability

QEDprocesses.compute

QEDprocesses.differential_cross_section

QEDprocesses.differential_cross_section

QEDprocesses.differential_probability

QEDprocesses.differential_probability

QEDprocesses.fundamental_interaction_type

QEDprocesses.in_phase_space_dimension

QEDprocesses.incoming_particles

QEDprocesses.number_incoming_particles

QEDprocesses.number_outgoing_particles

QEDprocesses.out_phase_space_dimension

QEDprocesses.outgoing_particles

QEDprocesses.physical_model

QEDprocesses.propagator

QEDprocesses.scattering_process

QEDprocesses.total_cross_section

QEDprocesses.total_cross_section

QEDprocesses.total_probability

QEDprocesses.total_probability

QEDprocesses.unsafe_differential_cross_section

QEDprocesses.unsafe_differential_cross_section

QEDprocesses.unsafe_differential_probability

QEDprocesses.unsafe_differential_probability

QEDprocesses.AbstractComputationSetup — Type

Abstract base type for computation setups. A setup means a collection of setup data needed to evaluate a dedicated quantity of given running data. Therefore, each setup is associated with a single quantity, which one may compute using the setup data and the running data. Despite that, the decomposition into setup and running data is arbitrary, and this can be used for cases where a subset of the variables a quantity depends on is kept constant.

Computation setup interface

The computation performed using a computation setup is separated into three steps:

1. input validation
2. actual computation
3. post processing

where every step has its own interface function (see compute for details).

Input validation

Every subtype of AbstractComputationSetup should implement the interface function

_assert_valid_input(stp::AbstractComputationSetup, input)

which should throw and an exception subtyped from AbstractInvalidInputException if the input is not valid for the computation of the associated quantity (see _assert_valid_input for more details). The default implementation does nothing, i.e. every input is valid by default. Provide a custom implementation if a different behavior is required.

Actual computation

Every subtype of AbstractComputationSetup must at least implement the required interface function

_compute(stp::AbstractComputationSetup, input) 

which computes the value of the associated quantity for a given input (see _compute for more details).

Post processing

Every subtype of AbstractComputationSetup should implement the interface function

_post_processing(stp::AbstractComputationSetup, input, result) 

which performs task after the actual computation, e.g. conversions or normalizations (see _post_processing for more details).

QEDprocesses.AbstractInvalidInputException — Type

Abstract base type for exceptions indicating invalid input. See InvalidInputError for a simple concrete implementation. Concrete implementations should at least implement

Base.showerror(io::IO, err::CustomInvalidError) where {CustomInvalidError<:AbstractInvalidInputException}

QEDprocesses.AbstractModelDefinition — Type

Abstract base type for all compute model definitions in the context of scattering processes. Every subtype of AbstractModelDefinition is associated with a fundamental interaction. Therefore, one needs to implement the following soft interface function

fundamental_interaction_type(::AbstractModelDefinition)

QEDprocesses.AbstractProcessDefinition — Type

Abstract base type for definitions of scattering processes. It is the root type for the process interface, which assumes that every subtype of AbstractProcessDefinition implements at least

incoming_particles(proc_def::AbstractProcessDefinition)
outgoing_particles(proc_def::AbstractProcessDefinition)

which return a tuple of the incoming and outgoing particles, respectively.

Furthermore, to calculate scattering probabilities and differential cross sections, the following interface functions need to be implemented

    _incident_flux(
        proc::AbstractProcessDefinition,
        model::AbstractModelDefinition, 
        in_phase_space::AbstractVector{T}
        ) where {T<:QEDbase.AbstractFourMomentum}

    _matrix_element(
        proc::AbstractProcessDefinition,
        model::AbstractModelDefinition, 
        in_phase_space::AbstractVector{T}
        out_phase_space::AbstractVector{T}
        ) where {T<:QEDbase.AbstractFourMomentum}

    _averaging_norm(
        proc::AbstractProcessDefinition
        )

    _is_in_phasespace(
        proc::AbstractProcessDefinition,
        model::AbstractModelDefinition, 
        in_ps_def::InPhasespaceDefinition,
        in_phase_space::AbstractVector{T}
        out_ps_def::OutPhasespaceDefinition,
        out_phase_space::AbstractVector{T}
    )

    _phase_space_factor(
        proc::AbstractProcessDefinition,
        model::AbstractModelDefinition, 
        in_ps_def::InPhasespaceDefinition,
        in_phase_space::AbstractVector{T}
        out_ps_def::OutPhasespaceDefinition,
        out_phase_space::AbstractVector{T}
        ) where {T<:QEDbase.AbstractFourMomentum}

Optional is the implementation of

    _total_probability(
        proc::AbstractProcessDefinition,
        model::AbstractModelDefinition, 
        in_ps_def::InPhasespaceDefinition,
        in_phase_space::AbstractVector{T}
    ) where {T<: QEDbase.AbstractFourMomentum}

to enable the calculation of total probabilities and cross sections.

QEDprocesses.AbstractProcessSetup — Type

Abstract base type for setups related to combining scattering processes and physical models. Every subtype of AbstractProcessSetup must implement at least the following interface functions:

scattering_process(::AbstractProcessSetup) 
physical_model(::AbstractProcessSetup) 

Derived from these interface functions, the following delegations are provided:

number_incoming_particles(::AbstractProcessSetup)
number_outgoing_particles(::AbstractProcessSetup)

QEDprocesses.InvalidInputError — Type

InvalidInputError(msg::String)

Exception which is thrown if a given input is invalid, e.g. passed to _assert_valid_input.

QEDprocesses.PhasespaceDefinition — Type

PhasespaceDefinition(coord_sys::AbstractCoordinateSystem, frame::AbstractFrameOfReference)

Convenient type to dispatch on coordiante systems and frames of reference.

QEDprocesses._assert_valid_input — Method

_assert_valid_input(stp::AbstractComputationSetup, input::Any)

Interface function, which asserts that the given input is valid, and throws an InvalidInputError if not.

default implementation

By default, every input is assumed to be valid. Therefore, this function does nothing. To customize this behavior, add your own implementation of

_assert_valid_input(stp::YourCustomSetup,input)

which should throw an exception, which is a subtype of AbstractInvalidInputException. One may also use the concrete implementation InvalidInputError if the input is invalid instead of writing a custom exception type.

QEDprocesses._averaging_norm — Function

_averaging_norm(
    proc::AbstractProcessDefinition
    )

Interface function, which returns a normalization for the averaging of the squared matrix elements over spins and polarizations.

QEDprocesses._base_component_type — Method

_base_component_type(array_of_lv::AbstractArray{LV}) where {LV<:QEDbase.AbstractLorentzVector}

Return the type of the components of given Lorentz vectors, which are by themself elements of an AbstractArray.

Examples

julia> using QEDbase
julia> using QEDprocesses
julia> v = Vector{SFourMomentum}(undef,10)
julia> QEDprocesses._base_component_type(v)
Float64

QEDprocesses._compute — Function

_compute(stp::AbstractComputationSetup, input::Any)

Interface function that returns the value of the associated quantity evaluated on input, which can be anything the associated quantity is defined to be feasible for.

unsafe implementation

This function must be implemented for any subtype of AbstractComputationSetup. It should not do any input validation or post processing (see _assert_valid_input and _post_processing), as those two are performed while calling the safe version of this function compute.

QEDprocesses._generate_incoming_momenta — Function

_generate_incoming_momenta
    proc::AbstractProcessDefinition,
    model::AbstractModelDefinition,
    in_phase_space_def::AbstractPhasespaceDefinition,
    in_phase_space::AbstractVector{T},
) where {T<:Real}

Interface function to generate the four-momenta of the incoming particles from coordinates for a given phase-space definition.

QEDprocesses._generate_momenta — Method

_generate_momenta(
proc::AbstractProcessDefinition,
model::AbstractModelDefinition,
in_phase_space_def::AbstractPhasespaceDefinition,
in_phase_space::AbstractVector{T},
out_phase_space_def::AbstractPhasespaceDefinition,
out_phase_space::AbstractVector{T},

) where {T<:Real}

Return four-momenta for incoming and outgoing particles for given coordinate based phase-space points.

QEDprocesses._generate_outgoing_momenta — Function

_generate_outgoing_momenta
    proc::AbstractProcessDefinition,
    model::AbstractModelDefinition,
    out_phase_space_def::AbstractPhasespaceDefinition,
    out_phase_space::AbstractVector{T},
) where {T<:Real}

Interface function to generate the four-momenta of the outgoing particles from coordinates for a given phase-space definition.

QEDprocesses._incident_flux — Function

_incident_flux(
    proc::AbstractProcessDefinition,
    model::AbstractModelDefinition, 
    in_phase_space::AbstractVector{T}
    ) where {T<:QEDbase.AbstractFourMomentum}

Interface function, which returns the incident flux of the given scattering process for a given incoming phase-space.

QEDprocesses._is_in_phasespace — Function

_is_in_phasespace(
    proc::AbstractProcessDefinition,
    model::AbstractModelDefinition, 
    in_ps_def::InPhasespaceDefinition,
    in_ps::AbstractVector{T}
    out_ps_def::OutPhasespaceDefinition,
    out_ps::AbstractVector{T}
    ) where {T<:QEDbase.AbstractFourMomentum}

Interface function, which returns true, if the combination of the given incoming and outgoing phase space is physical, i.e. all momenta are on-shell and some sort of energy-momentum conservation holds.

QEDprocesses._matrix_element — Function

_matrix_element(
    proc::AbstractProcessDefinition,
    model::AbstractModelDefinition, 
    in_phase_space::AbstractVector{T}
    out_phase_space::AbstractVector{T}
    ) where {T<:QEDbase.AbstractFourMomentum}

Interface function, which returns a tuple of scattering matrix elements for each spin and polarisation combination of proc.

QEDprocesses._phase_space_factor — Function

_phase_space_factor(
    proc::AbstractProcessDefinition,
    model::AbstractModelDefinition, 
    in_ps_def::InPhasespaceDefinition,
    in_phase_space::AbstractVector{T}
    out_ps_def::OutPhasespaceDefinition,
    out_phase_space::AbstractVector{T}
    ) where {T<:QEDbase.AbstractFourMomentum}

Interface function, which returns the pre-differential factor of the invariant phase space intergral measure.

Convention

It is assumed, that this function returns the value of

\[\mathrm{d}\Pi_n:= \prod_{i=1}^N \frac{\mathrm{d}^3p_i}{(2\pi)^3 2 p_i^0} H(P_t, p_1, \dots, p_N),\]

where $H(\dots)$ is a characteristic function (or distribution) which constrains the phase space, e.g. $\delta^{(4)}(P_t - \sum_i p_i)$.

QEDprocesses._post_processing — Method

function _post_processing(stp::AbstractComputationSetup, input::Any, result::Any)

Interface function, which is called in compute after _compute has been called. This function is dedicated to finalize the result of a computation.

default implementation

Since in the case of no post processing the result of _compute is unchanged, this function returns result by default.

QEDprocesses._total_probability — Function

_total_probability(
    proc::AbstractProcessDefinition,
    model::AbstractModelDefinition, 
    in_ps_def::InPhasespaceDefinition,
    in_phase_space::AbstractVector{T}
) where {T<: QEDbase.AbstractFourMomentum}

Interface function for the combination of a scattering process and a physical model. Return the total of a given process and model for a passed initial phase space definition and point. The elements of in_phase_space, which represent the initial phase space, are the momenta of the respective particles. The implementation of this function for a concrete process and model must not check if the length of the passed initial phase spaces match the number of incoming particles.

probability interface

Given an implementation of this method, the following generic implementation without input check is provided:

_total_probability(proc_def,model_def,in_phasespace::AbstractMatrix{T})

where T<:QEDbase.AbstractFourMomentum, i.e. _total_probability is also implemented for a vector of initial phase space points. Furthermore, a safe version of _total_probability is also implemented: total_probability.

total cross section

Given an implementaion of this method and _incident_flux, the respective functions for the total cross section are available, i.e. _total_cross_section (unsafe and not exported), and total_cross_section, respectively.

Note

Each instance of this function does not check the validity of the input. This function is not exported and should be used with caution. To add a method in order to implement the cross section interface, it is recommented to directly use QEDprocesses._total_cross_section instead of globally using QEDprocesses: _total_cross_section.

QEDprocesses.compute — Method

compute(stp::AbstractComputationSetup, input::Any)

Return the value of the quantity associated with stp for a given input. In addition to the actual call of the associated unsafe version _compute, input validation ([_assert_valid_input]) and post processing (using _post_processing) are wrapped around the calculation (see AbstractComputationSetup for details).

QEDprocesses.differential_cross_section — Method

differential_cross_section(
    proc::AbstractProcessDefinition,
    model::AbstractModelDefinition,
    in_phase_space_def::AbstractPhasespaceDefinition,
    in_phase_space::AbstractVecOrMat{T},
    out_phase_space_def::AbstractPhasespaceDefinition,
    out_phase_space::AbstractVecOrMat{T},
) where {T<:QEDbase.AbstractFourMomentum}

If the given phase spaces are physical, return differential cross section evaluated at the four-momenta. Zero otherwise.

QEDprocesses.differential_cross_section — Method

differential_cross_section(
proc::AbstractProcessDefinition,
model::AbstractModelDefinition,
in_phase_space_def::AbstractPhasespaceDefinition,
in_phase_space::AbstractVecOrMat{T},
out_phase_space_def::AbstractPhasespaceDefinition,
out_phase_space::AbstractVecOrMat{T},

) where {T<:Real}

If the given phase spaces are physical, return differential cross section evaluated at the coordinates. Zero otherwise.

QEDprocesses.differential_probability — Method

differential_probability(
    proc::AbstractProcessDefinition,
    model::AbstractModelDefinition,
    in_phase_space_def::AbstractPhasespaceDefinition,
    in_phase_space::AbstractVecOrMat{T},
    out_phase_space_def::AbstractPhasespaceDefinition,
    out_phase_space::AbstractVecOrMat{T},
) where {T<:QEDbase.AbstractFourMomentum}

If the given phase spaces are physical, return differential probability evaluated at the four-momenta. Zero otherwise.

QEDprocesses.differential_probability — Method

differential_probability(
proc::AbstractProcessDefinition,
model::AbstractModelDefinition,
in_phase_space_def::AbstractPhasespaceDefinition,
in_phase_space::AbstractVecOrMat{T},
out_phase_space_def::AbstractPhasespaceDefinition,
out_phase_space::AbstractVecOrMat{T},

) where {T<:Real}

If the given phase spaces are physical, return differential probability evaluated at the coordinates. Zero otherwise.

QEDprocesses.fundamental_interaction_type — Function

fundamental_interaction_type(models_def::AbstractModelDefinition)

Return the fundamental interaction associated with the passed model definition.

QEDprocesses.in_phase_space_dimension — Function

in_phase_space_dimension(
    proc::AbstractProcessDefinition,
    model::AbstractModelDefinition,
    )

TBW

QEDprocesses.incoming_particles — Function

incoming_particles(proc_def::AbstractProcessDefinition)

Interface function for scattering processes. Return a tuple of the incoming particles for the given process definition. This function needs to be given to implement the scattering process interface.

QEDprocesses.number_incoming_particles — Method

number_incoming_particles(proc_def::AbstractProcessDefinition)

Return the number of incoming particles of a given process.

QEDprocesses.number_outgoing_particles — Method

number_outgoing_particles(proc_def::AbstractProcessDefinition)

Return the number of outgoing particles of a given process.

QEDprocesses.out_phase_space_dimension — Function

out_phase_space_dimension(
    proc::AbstractProcessDefinition,
    model::AbstractModelDefinition,
    )

TBW

QEDprocesses.outgoing_particles — Function

outgoing_particles(proc_def::AbstractProcessDefinition)

Interface function for scattering processes. Return the tuple of outgoing particles for the given process definition. This function needs to be given to implement the scattering process interface.

QEDprocesses.physical_model — Function

physical_model(stp::AbstractProcessSetup)

Interface function that returns the physical model associated with stp, i.e. an object which is a subtype of AbstractModelDefinition.

QEDprocesses.propagator — Function

propagator(particle::AbstractParticleType, mom::QEDbase.AbstractFourMomentum, [mass::Real])

Return the propagator of a particle for a given four-momentum. If mass is passed, the respective propagator for massive particles is used, if not, it is assumed the particle passed in is massless.

Convention

There are two types of implementations for propagators given in QEDProcesses: For a BosonLike particle with four-momentum $k$ and mass $m$, the propagator is given as

\[D(k) = \frac{1}{k^2 - m^2}.\]

For a FermionLike particle with four-momentum $p$ and mass $m$, the propagator is given as

\[S(p) = \frac{\gamma^\mu p_\mu + mass}{p^2 - m^2}.\]

Warning

This function does not throw when the given particle is off-shell. If an off-shell particle is passed, the function propagator returns Inf.

QEDprocesses.scattering_process — Function

scattering_process(stp::AbstractProcessSetup)

Interface function that returns the scattering process associated with stp, i.e. an object which is a subtype of AbstractProcessDefinition.

QEDprocesses.total_cross_section — Method

total_cross_section(
    proc::AbstractProcessDefinition,
    model::AbstractModelDefinition,
    in_phase_space_def::AbstractPhasespaceDefinition,
    in_phase_space::AbstractVecOrMat{T},
) where {T<:QEDbase.AbstractFourMomentum}

Return the total cross section for a given combination of scattering process and compute model, evaluated at the particle momenta.

QEDprocesses.total_cross_section — Method

total_cross_section(
    proc::AbstractProcessDefinition,
    model::AbstractModelDefinition,
    in_phase_space_def::AbstractPhasespaceDefinition,
    in_phase_space::AbstractVecOrMat{T},
) where {T<:Real}

Return the total cross section for a given combination of scattering process and compute model, evaluated at the coordinates.

QEDprocesses.total_probability — Method

total_probability(
    proc::AbstractProcessDefinition,
    model::AbstractModelDefinition,
    in_phase_space_def::AbstractPhasespaceDefinition,
    in_phase_space::AbstractMatrix{T},
) where {T<:QEDbase.AbstractFourMomentum}

Return the total probability of a given model and process combination, evaluated at the particle momenta.

QEDprocesses.total_probability — Method

total_probability(
    proc::AbstractProcessDefinition,
    model::AbstractModelDefinition,
    in_phase_space_def::AbstractPhasespaceDefinition,
    in_phase_space::AbstractMatrix{T},
) where {T<:Real}

Return the total probability of a given model and process combination, evaluated at the coordinates.

QEDprocesses.unsafe_differential_cross_section — Method

function unsafe_differential_cross_section(
    proc::AbstractProcessDefinition,
    model::AbstractModelDefinition,
    in_phase_space_def::AbstractPhasespaceDefinition,
    in_phase_space::AbstractVecOrMat{T},
    out_phase_space_def::AbstractPhasespaceDefinition,
    out_phase_space::AbstractVecOrMat{T},
) where {T<:QEDbase.AbstractFourMomentum}

Return the differential cross section evaluated at the four-momenta without checking if the given phase space(s) are physical.

QEDprocesses.unsafe_differential_cross_section — Method

unsafe_differential_cross_section(
proc::AbstractProcessDefinition,
model::AbstractModelDefinition,
in_phase_space_def::AbstractPhasespaceDefinition,
in_phase_space::AbstractVecOrMat{T},
out_phase_space_def::AbstractPhasespaceDefinition,
out_phase_space::AbstractVecOrMat{T},

) where {T<:Real}

Return the differential cross section evaluated at the coordinates without checking if the given phase space(s) are physical.

QEDprocesses.unsafe_differential_probability — Method

unsafe_differential_probability(
    proc::AbstractProcessDefinition,
    model::AbstractModelDefinition,
    in_phase_space_def::AbstractPhasespaceDefinition,
    in_phase_space::AbstractVecOrMat{T},
    out_phase_space_def::AbstractPhasespaceDefinition,
    out_phase_space::AbstractVecOrMat{T},
) where {T<:QEDbase.AbstractFourMomentum}

Return differential probability evaluated at the four-momenta without checking if the given phase space(s) are physical.

QEDprocesses.unsafe_differential_probability — Method

unsafe_differential_probability(
    proc::AbstractProcessDefinition,
    model::AbstractModelDefinition,
    in_phase_space_def::AbstractPhasespaceDefinition,
    in_phase_space::AbstractVecOrMat{T},
    out_phase_space_def::AbstractPhasespaceDefinition,
    out_phase_space::AbstractVecOrMat{T},
) where {T<:Real}

Return differential probability evaluated at the coordinates without checking if the given phase space(s) are physical.